+++ bpm-engine-api/src/main/java/io/takari/bpm/model/ServiceTask.java
@@ -1,5 +1,7 @@
 package io.takari.bpm.model;
 
+import java.util.Set;
+
 public class ServiceTask extends AbstractElement {
 	
 	private static final long serialVersionUID = 1L;
@@ -7,15 +9,23 @@ public class ServiceTask extends AbstractElement {
     private String name;
     private final String expression;
     private final ExpressionType type;
+    private final Set<VariableMapping> in;
+    private final Set<VariableMapping> out;
 
     public ServiceTask(String id) {
         this(id, ExpressionType.NONE, null);
     }
 
     public ServiceTask(String id, ExpressionType type, String expression) {
+        this(id, type, expression, null, null);
+    }
+
+    public ServiceTask(String id, ExpressionType type, String expression, Set<VariableMapping> in, Set<VariableMapping> out) {
         super(id);
         this.type = type;
         this.expression = expression;
+        this.in = in;
+        this.out = out;
     }
 
     public String getExpression() {
@@ -32,5 +42,13 @@ public class ServiceTask extends AbstractElement {
 
     public void setName(String name) {
         this.name = name;
-    }            
+    }
+
+    public Set<VariableMapping> getIn() {
+        return in;
+    }
+
+    public Set<VariableMapping> getOut() {
+        return out;
+    }
 }